requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.35.0
httpx==0.27.0
pydantic==2.7.4
python-multipart==0.0.9
aiofiles==23.2.1
Pillow==10.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
uvicorn[standard]
APScheduler==3.10.4
slowapi==0.1.9
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
structlog==24.1.0
sentry-sdk[fastapi]==2.14.0
asyncpg==0.29.0
python-dateutil==2.9.0
pytest==8.3.2
pytest-asyncio==0.23.8
redis>=5.0.0
cryptography>=42.0.0
